requirements.txt
fake_headers>=1.0.2
pandas>=2.0.3
python-dotenv>=1.0.0
selenium>=4.12.0
webdriver_manager>=4.0.0
langchain>=0.0.142
langgraph>=0.0.1
openai>=0.27.0
wikipedia>=1.4.0
langchain-openai>=0.3.7
